feat: isolate providers from layout as components
