+++ src/layouts/Main.tsx
@@ -1,15 +1,9 @@
 import React from "react";
-import {
-  createStyles,
-  makeStyles,
-  Paper,
-  ThemeProvider,
-} from "@material-ui/core";
+import { createStyles, makeStyles, Paper } from "@material-ui/core";
 import TopBar from "@/components/TopBar";
 import BottomNav from "@/components/BottomNav";
-import { TitleContextProvider } from "@/libs/title";
-import { ThemeContextProvider, useTheme } from "@/libs/themeMode";
-import { useHistory } from "react-router-dom";
+import { useTitleContext } from "@/libs/title";
+import ProvidersProvider from "@/components/ProvidersProvider";
 
 const useStyles = makeStyles(() =>
   createStyles({
@@ -35,54 +29,27 @@ const useStyles = makeStyles(() =>
   })
 );
 
-interface Props {
-  children: React.ReactNode;
-}
-const TITLE_SUFFIX = "73rd Afes Manage App";
-const TOP_TITLE = "73rd Afes Manage App";
-
-const MainLayout: React.FC<Props> = (props) => {
+const MainLayout: React.FC = (props) => {
   const classes = useStyles();
-  const history = useHistory();
-  const [titleState, _setTitleState] = React.useState({
-    title: "",
-    _setTitle,
-  });
-
-  function _setTitle(_new_title: string) {
-    _setTitleState((old) => ({ ...old, title: _new_title }));
-    document.title = _new_title + " - " + TITLE_SUFFIX;
-    if (_new_title === "" || history.location.pathname === "/")
-      document.title = TOP_TITLE;
-  }
+  const titleCtx = useTitleContext();
 
   return (
-    <ToggleThemeProvider>
-      <TitleContextProvider value={titleState}>
-        <Paper className={classes.root} square={true}>
-          <div className={classes.topBar}>
-            <TopBar title={titleState.title} />
-          </div>
-          <main className={classes.main}>{props.children}</main>
-          <div className={classes.bottomNav}>
-            <BottomNav />
-          </div>
-        </Paper>
-      </TitleContextProvider>
-    </ToggleThemeProvider>
+    <Paper className={classes.root} square={true}>
+      <div className={classes.topBar}>
+        <TopBar title={titleCtx.title} />
+      </div>
+      <main className={classes.main}>{props.children}</main>
+      <div className={classes.bottomNav}>
+        <BottomNav />
+      </div>
+    </Paper>
   );
 };
 
-export default MainLayout;
+const MainLayoutWithProviders: React.FC = (props) => (
+  <ProvidersProvider>
+    <MainLayout>{props.children}</MainLayout>
+  </ProvidersProvider>
+);
 
-const ToggleThemeProvider: React.FC<{
-  children: React.ReactNode;
-}> = (props) => {
-  const [theme, setThemeMode] = useTheme();
-
-  return (
-    <ThemeContextProvider value={{ setThemeMode }}>
-      <ThemeProvider theme={theme}>{props.children}</ThemeProvider>
-    </ThemeContextProvider>
-  );
-};
+export default MainLayoutWithProviders;
